Add coffee-script dependency. Remove compiled js.
I added coffee-script as a devDependency, which makes it so anyone
trying to work on the module gets coffee-script. I also added a
prepublish script to automatically build the coffee files when you go to
publish the package. I also added a build script so you can use "npm
run-script build" to regenerate the .js files.

Since the .js files are really build artifacts, they shouldn't be
checked in to git. I deleted the .js files, and added them to
.gitignore. Also, I set up a .npmignore so the .coffee files and the
tests don't get included in the npm module.

From what I've googled, this seems to be the current best practice. I
took the examples from the [zombie
module](https://github.com/assaf/zombie), which seems to be a popular
node module written in CoffeeScript.